# backend/src/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Bitácora SOC API
  description: API completa para el sistema de Bitácora SOC con gestión de entradas, checklist de turno, usuarios y reportes
  version: 1.0.0
  contact:
    name: Bitácora SOC Team
servers:
  - url: http://192.168.100.50:3000/api
    description: Servidor de desarrollo (IP)
  - url: http://localhost:3000/api
    description: Servidor local
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ShiftCheck:
      type: object
      properties:
        _id:
          type: string
        userId:
          type: string
        username:
          type: string
        type:
          type: string
          enum: [inicio, cierre]
        checkDate:
          type: string
          format: date-time
        services:
          type: array
          items:
            type: object
            properties:
              serviceId:
                type: string
              serviceTitle:
                type: string
              status:
                type: string
                enum: [verde, rojo]
              observation:
                type: string
        hasRedServices:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
paths:
...
  /checklist/services:
    get:
      tags:
        - Checklist
      summary: Obtener servicios activos
      description: Lista los servicios activos del catálogo
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de servicios
    
    post:
      tags:
        - Checklist
      summary: Crear servicio
      description: Crea un nuevo servicio en el catálogo (solo admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
              properties:
                title:
                  type: string
                  example: QRadar
                order:
                  type: integer
                  example: 1
      responses:
        '201':
          description: Servicio creado
  /checklist/check:
    post:
      tags:
        - Checklist
      summary: Registrar checklist de turno
      description: |
        Registra un checklist de inicio o cierre de turno
        
        **Reglas:**
        - No se permiten dos acciones iguales consecutivas
        - Debe respetar el cooldown configurable (default 4 horas)
        - Todos los servicios en rojo DEBEN tener observación
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - type
                - services
              properties:
                type:
                  type: string
                  enum: [inicio, cierre]
                  example: inicio
                services:
                  type: array
                  items:
                    type: object
                    required:
                      - serviceId
                      - serviceTitle
                      - status
                    properties:
                      serviceId:
                        type: string
                      serviceTitle:
                        type: string
                      status:
                        type: string
                        enum: [verde, rojo]
                      observation:
                        type: string
                        maxLength: 1000
            example:
              type: inicio
              services:
                - serviceId: "675e1234567890abcdef1234"
                  serviceTitle: QRadar
                  status: verde
                  observation: ""
                - serviceId: "675e1234567890abcdef1235"
                  serviceTitle: Zabbix
                  status: rojo
                  observation: "Alerta de CPU en servidor prod-01"
      responses:
        '201':
          description: Checklist registrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  check:
                    $ref: '#/components/schemas/ShiftCheck'
        '400':
          description: Error de validación o reglas de negocio
  /checklist/check/last:
    get:
      tags:
        - Checklist
      summary: Último check del usuario
      description: Obtiene el último checklist registrado por el usuario
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Último check
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShiftCheck'
...
  /backup/mongo:
    get:
      tags:
        - Backup
      summary: Crear backup de MongoDB
      description: Crea un backup completo de la base de datos (solo admin)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Backup creado
  /backup/list:
    get:
      tags:
        - Backup
      summary: Listar backups
      description: Lista todos los backups disponibles (solo admin)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de backups
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    name:
                      type: string
                    path:
                      type: string
                    created:
                      type: string
                      format: date-time
                    size:
                      type: integer